Re-organized files and started setup to work on Create account page
